The way Scintilla makes the selected item appear in the middle of the list (and away from the caret and focus of the user) instead of at the top does not make a lot of sense to me, especially because it even does this when the items above the selection have no match of any kind and are completely pointless to display.
In VS the items above the selected one are only the partially matching ones (with this I mean: ABC partially matches XAYBZC because it has all letters in order). If there aren't any partial matches (like in Scintilla which doesn't do these: it matches ABC to ABCXYZ but not to XAYBZC) then the selection is at the top.

VSCode improves on this: it puts the partial matches below the regular matches. So again the selection and best match is at the top, close to the caret and focus.

Ideally there would be more Scintilla changes: it should try more variants of matchings and put those below the regular matches and should remove anything which does not match (even Delphi which is pretty bad at all of this does the latter).